+++ src/server/middleware/bundling-check.js
@@ -1,0 +1,13 @@
+import Layout from '../layout-bundling';
+
+/**
+ * @param {import('..').default} server
+ */
+export default (server) => (req, res, next) => {
+	if (!server.bundling) {
+		next();
+		return;
+	}
+	res.header('content-type', 'text/html');
+	res.renderLayout({ layout: Layout });
+};